chore: invert the parameters type dependency
